adding a script in order to check all projects.
When I modify something on libC, it is very annoying to check if a
project has been broken by the new modification.
Add a script that will do that for me.
It is possible with this script to compile, run or clean the projects.
It is also possible to filter on a specify project.