p67: add resolution of the problem p67.
I had already resolved this problem but I did not pushed the code. Now,
it is done.

Note: the code is exactly the same than p18 because it is the same
problem with much more data in this problem. But, my algorithm is p18
was already sufficient to resolve the problem p67 efficiently.