pz136: add the resolution of this problem.
This problem is very similar to the 135 one.
The code is almost identical. I have decided to not share the code
because it is too more specific according to me to add it in libC.
I could add the code in a 'algo.c' file in pz135 and then link this file
in pz136, but it would have not necessarily better, so, just keep it like
that for now.